refactor: update packages
