+++ app/src/main/java/com/notescomposeapp/ui/views/CardNote.kt
@@ -1,11 +1,10 @@
-package com.notescomposeapp.ui.screens
+package com.notescomposeapp.ui.views
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
-import androidx.compose.material.Shapes
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
